requirements.txt
ruamel.yaml
Jinja2
requests
